+++ tests/content-middleware.js
@@ -3,7 +3,8 @@ import uuid from 'uuid';
 import cloneDeep from 'lodash/cloneDeep';
 
 import types from './fixtures/content-types/objects/all-merged';
-import utils from '../lib/content/utils';
+import middleware from '../lib/content/middleware';
+import utils from '../lib/content/middleware/utils';
 
 /**
  * Next
@@ -197,7 +198,7 @@ test('URL checks type exists in CMS', t => {
   const rq = cloneDeep(req);
   rq.params.type = 'foo';
 
-  return utils.url(rq, {}, next).then(err => {
+  return middleware.url(rq, {}, next).then(err => {
     t.is(err.message, 'Content Type \'foo\' not found', 'Should fail with error object');
     t.is(err.safe, '/content', 'Should have a safe url');
     t.is(err.status, 404, 'Should be a 404');
@@ -208,7 +209,7 @@ test('URL checks id is uuid', t => {
   const rq = cloneDeep(req);
   rq.params.id = 'foo';
 
-  return utils.url(rq, {}, next).then(err => {
+  return middleware.url(rq, {}, next).then(err => {
     t.is(err.message, 'Content ID must be in UUID format', 'Should fail with message');
     t.is(err.safe, '/content', 'Should have a safe url');
     t.is(err.status, 404, 'Should be a 404');
@@ -220,7 +221,7 @@ test('URL checks revision is number', t => {
   rq.params.id = uuid.v4();
   rq.params.revision = 'foo';
 
-  return utils.url(rq, {}, next).then(err => {
+  return middleware.url(rq, {}, next).then(err => {
     t.is(err.message, 'Revision must be a number', 'Should fail with message');
     t.is(err.safe, '/content', 'Should have a safe url');
     t.is(err.status, 404, 'Should be a 404');
@@ -232,7 +233,7 @@ test('URL works with CMS as it is configured', t => {
   rq2.params.id = uuid.v4();
   rq2.params.revision = 2345;
 
-  return utils.url(rq2, {}, next).then(() => {
+  return middleware.url(rq2, {}, next).then(() => {
     t.pass();
   }).catch(() => {
     t.fail();
